agent_okto_3/requirements.txt
# Principais bibliotecas
faiss-cpu==1.8.0
openai==1.13.3
python-dotenv==1.0.1
langchain==0.2.16
langchainhub==0.1.15
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-experimental==0.0.65
langchain-community==0.2.16
langchain-huggingface==0.1.0
chromadb==0.5.3
pydantic==2.9.1
pandas==2.2.2
streamlit==1.38.0
sentence-transformers==3.0.1

# Dependências auxiliares
aiohttp==3.10.5
altair==5.4.1
anyio==4.4.0
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
fastapi==0.114.0
fsspec==2024.9.0
gitpython==3.1.43
grpcio==1.66.1
huggingface-hub==0.24.6
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
numpy==1.26.4
pillow==10.4.0
protobuf==4.25.4
pytz==2024.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
scikit-learn==1.5.1
scipy==1.14.1
sqlalchemy==2.0.34
tqdm==4.66.5
transformers==4.44.2
uvicorn==0.30.6
websockets==13.0.1
